requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
Pillow==9.0.1
scipy
ipython
jupyter
pandas
sympy
nose
plotly
scikit-learn
opencv-python
open3d